refactor: update User type import paths across admin components
