+++ src/components/admin/UserFormModal.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import type { User } from '../../services/api';
+import type { User } from '../../types/auth';
 
 interface UserFormModalProps {
   isOpen: boolean;
