+++ src/components/admin/UsersTable.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { UserRow } from './UserRow';
-import type { User } from '../../services/api';
+import type { User } from '../../types/auth';
 import { USER_MESSAGES } from '../../constants/admin/userMessages';
 import { getRoleInfo } from '../../constants/admin/userRoles';
